pyproject.toml
[tool.poetry]
name = "qap-final-project"
version = "0.1.0"
description = "Skillfactory QAP final project"
authors = ["Mikhail Lyalin <mc.lyalin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"


[tool.poetry.group.dev.dependencies]
pytest-selenium = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
